manifold runner: avoid defmethod for task dispatch
In the manifold runner for ablauf workflows, actions are dispatched
on with the help of the `ablauf.job.manifold/dispatch-action`
multi-method.
It is common for programs to use system wiring libraries such
as component or integrant to perform initialization. Using a
multimethod makes it harder than necessary to integrate with
these systems.

This patch switches to using a provided `action-fn`, leaving
it to the library consumer to use the best possible approach
for dispatching.