proof of concept sql runner
A minimalist job queue worker for SQL. This assumes
multiple worker off of a main job queue. In the following
code, each worker is assumed to be able to process both
actions and workflow evaluation. The processing is based off
of two tables: `task` and `workflow_run`. `workflow_run` holds
information about a full workflow run and its current status,
while `task` is used for both calls to `job/restart` and action
dispatches.